# Conflicts:
#	.gradle/7.4/executionHistory/executionHistory.bin
#	.gradle/7.4/executionHistory/executionHistory.lock
#	.gradle/7.4/fileHashes/fileHashes.bin
#	.gradle/7.4/fileHashes/fileHashes.lock
#	.gradle/7.4/fileHashes/resourceHashesCache.bin
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock

+++ app/src/main/java/classes/Dinner.java
@@ -333,7 +333,7 @@ public class Dinner {
     }
 
     public static void deletePicture(String picName){
-        if(!picName.equals("default_dinner.jpg")) {
+        if(!picName.equals("default_dinner.jpg")&&!picName.equals("")) {
             StorageReference storageReference = FirebaseStorage.getInstance().getReference("images/" + picName);
             storageReference.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
                 @Override
